레포츠 평점 출력 수정

+++ src/main/java/com/dto/LeportsDetailDTO.java
@@ -34,5 +34,5 @@ public class LeportsDetailDTO {
 	private String c_phone;
 	private String leports_content;
 	private int review_cnt;
-	private int review_star;
+	private double review_star;
 }
